Reduce draw calls and optimize lighting further
