+++ Protogame/Core/ProtogameCoreModule.cs
@@ -58,7 +58,9 @@ namespace Protogame
             kernel.Bind<ILightFactory>().ToFactory().DiscardNodeOnResolve();
             kernel.Bind<IStandardDirectionalLight>().To<DefaultStandardDirectionalLight>().InParentScope().AllowManyPerScope();
             kernel.Bind<IStandardPointLight>().To<DefaultStandardPointLight>().InParentScope().AllowManyPerScope();
-            
+            kernel.Bind<ILightRenderer<IStandardDirectionalLight>>().To<DefaultStandardDirectionalLightRenderer>().InSingletonScope();
+            kernel.Bind<ILightRenderer<IStandardPointLight>>().To<DefaultStandardPointLightRenderer>().InSingletonScope();
+
             kernel.Bind<IDebugRenderer>().To<NullDebugRenderer>().InSingletonScope();
 
             kernel.Bind<IRenderBatcher>().To<DefaultRenderBatcher>().InSingletonScope();
