+++ Protogame/Graphics/Lighting/ILight.cs
@@ -2,6 +2,6 @@
 {
     public interface ILight
     {
-        void Render(IGameContext gameContext, IRenderContext renderContext, ILightContext lightContext);
+        ILightRenderer LightRenderer { get; }
     }
 }
