+++ Protogame/Graphics/Lighting/ILightRenderer.cs
@@ -1,0 +1,13 @@
+﻿using System.Collections.Generic;
+
+namespace Protogame
+{
+    public interface ILightRenderer<T> : ILightRenderer
+    {
+    }
+
+    public interface ILightRenderer
+    {
+        void Render(IGameContext gameContext, IRenderContext renderContext, ILightContext lightContext, IEnumerable<ILight> lights);
+    }
+}
